Using o dbus format for bus name, and u dbus format for account id. Renamed AccountDetails members and Account methods to be clearer.

+++ include/unity/storage/qt/internal/RuntimeImpl.h
@@ -61,7 +61,7 @@ public:
 
     Account make_test_account(QString const& bus_name,
                               QString const& object_path,
-                              qlonglong id,
+                              quint32 id,
                               QString const& service_id,
                               QString const& display_name);
